Open Source
http://freedoms.land/source.html

+++ Potatso/Base/QRCode/HMScannerBorder.m
@@ -62,7 +62,7 @@
     
     // 加载边框图像
     for (NSInteger i = 1; i < 5; i++) {
-        NSString *imgName = [NSString stringWithFormat:@"ScanQR%zd", i];
+        NSString *imgName = [NSString stringWithFormat:@"ScanQR%ld", (long)i];
         UIImageView *img = [[UIImageView alloc] initWithImage:[self imageWithName:imgName bundle:imageBundle]];
         
         [self addSubview:img];
